+++ lib/index.ts
@@ -1,5 +1,6 @@
 import * as crypto from 'crypto';
 import request from 'superagent';
+import queystring from 'querystring';
 
 export interface MemberDetails {
   email_address?: string;
@@ -60,6 +61,13 @@ export class Mailchimp {
     const emailHash = Mailchimp.getEmailHash(email);
     const path = `/lists/${listId}/members/${emailHash}`;
     return this.sendRequest('DELETE', path);
+  }
+
+  /** https://mailchimp.com/developer/marketing/api/list-members/list-members-info/ */
+  getMembers(listId: string, options: { count?: number; offset?: number } = {}): Promise<any> {
+    const query = queystring.stringify(options);
+    const path = `/lists/${listId}/members?${query}`;
+    return this.sendRequest('GET', path);
   }
 
   private sendRequest(method: string, path: string, body?: any): Promise<any> {